move things into /admin/models

+++ src/pages/admin/modelConfig/components/ModelConfigurationList/deleteModelAction.ts
@@ -14,7 +14,7 @@ export const deleteModelAction =
             throw Error('Model Id is required!');
         }
 
-        await playgroundApiClient.DELETE('/v4/models/{model_id}', {
+        await playgroundApiClient.DELETE('/v4/admin/models/{model_id}', {
             params: {
                 path: { model_id: modelId },
             },
